appserver1/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.8.6
redis==3.5.3